Move bloom filter definitions from old file to new and fix includes
