+++ src/include/duckdb/planner/filter/bloom_filter.hpp
@@ -11,36 +11,11 @@
 #include "duckdb/common/optional_ptr.hpp"
 #include "duckdb/common/serializer/deserializer.hpp"
 #include "duckdb/common/serializer/serializer.hpp"
-#include "duckdb/common/types/value.hpp"
 #include "duckdb/planner/table_filter.hpp"
-#include "duckdb/storage/buffer_manager.hpp"
 
 namespace duckdb {
 
-//! Runtime bloom-filter state used by join pushdown and internal tablefilter functions.
-class BloomFilter {
-public:
-	BloomFilter() = default;
-	void Initialize(ClientContext &context_p, idx_t number_of_rows);
-
-	void InsertHashes(const Vector &hashes_v, idx_t count) const;
-	idx_t LookupHashes(const Vector &hashes_v, SelectionVector &result_sel, idx_t count) const;
-
-	void InsertOne(hash_t hash) const;
-	bool LookupOne(hash_t hash) const;
-
-	bool IsInitialized() const {
-		return initialized;
-	}
-
-private:
-	idx_t num_sectors;
-	uint64_t bitmask; // num_sectors - 1 -> used to get the sector offset
-
-	bool initialized = false;
-	AllocatedData buf_;
-	uint64_t *bf;
-};
+class BloomFilter;
 
 //! DEPRECATED - only preserved for backwards-compatible expression conversion
 class LegacyBFTableFilter final : public TableFilter {
