+++ src/include/duckdb/planner/filter/table_filter_functions.hpp
@@ -21,7 +21,6 @@ namespace duckdb {
 class BaseStatistics;
 class Expression;
 class PerfectHashJoinExecutor;
-class BloomFilter;
 class PrefixRangeFilter;
 struct DynamicFilterData;
 enum class SelectivityOptionalFilterType : uint8_t;
@@ -44,6 +43,31 @@ struct TableFilterFunctions {
 	static bool IsTableFilterFunction(const BoundScalarFunction &function) {
 		return IsTableFilterFunction(function.GetName());
 	}
+};
+
+//! Runtime bloom-filter state used by join pushdown and internal tablefilter functions.
+class BloomFilter {
+public:
+	BloomFilter() = default;
+	void Initialize(ClientContext &context_p, idx_t number_of_rows);
+
+	void InsertHashes(const Vector &hashes_v, idx_t count) const;
+	idx_t LookupHashes(const Vector &hashes_v, SelectionVector &result_sel, idx_t count) const;
+
+	void InsertOne(hash_t hash) const;
+	bool LookupOne(hash_t hash) const;
+
+	bool IsInitialized() const {
+		return initialized;
+	}
+
+private:
+	idx_t num_sectors;
+	uint64_t bitmask; // num_sectors - 1 -> used to get the sector offset
+
+	bool initialized = false;
+	AllocatedData buf_;
+	uint64_t *bf;
 };
 
 //! FunctionData for bloom filter internal function
