+++ src/planner/filter/table_filter_bloom_function.cpp
@@ -8,7 +8,6 @@
 
 #include "duckdb/planner/filter/table_filter_functions.hpp"
 #include "duckdb/planner/filter/table_filter_function_helpers.hpp"
-#include "duckdb/planner/filter/bloom_filter.hpp"
 
 #include "duckdb/common/operator/cast_operators.hpp"
 #include "duckdb/common/operator/subtract.hpp"
@@ -17,8 +16,84 @@
 #include "duckdb/execution/expression_executor_state.hpp"
 #include "duckdb/planner/expression/bound_function_expression.hpp"
 #include "duckdb/storage/statistics/numeric_stats.hpp"
+#include "duckdb/storage/buffer_manager.hpp"
 
 namespace duckdb {
+
+static constexpr idx_t MAX_NUM_SECTORS = (1ULL << 26);
+static constexpr idx_t MIN_NUM_BITS_PER_KEY = 12;
+static constexpr idx_t MIN_NUM_BITS = 512;
+static constexpr idx_t LOG_SECTOR_SIZE = 6;             // a sector is 64 bits, log2(64) = 6
+static constexpr idx_t SHIFT_MASK = 0x3F3F3F3F3F3F3F3F; // 6 bits for 64 positions
+static constexpr idx_t N_BITS = 4;                      // the number of bits to set per hash
+
+void BloomFilter::Initialize(ClientContext &context_p, idx_t number_of_rows) {
+	BufferManager &buffer_manager = BufferManager::GetBufferManager(context_p);
+
+	const idx_t min_bits = MaxValue(MIN_NUM_BITS, number_of_rows * MIN_NUM_BITS_PER_KEY);
+	num_sectors = MinValue(NextPowerOfTwo(min_bits) >> LOG_SECTOR_SIZE, MAX_NUM_SECTORS);
+	bitmask = num_sectors - 1;
+
+	buf_ = buffer_manager.GetBufferAllocator().Allocate(64 + num_sectors * sizeof(uint64_t));
+	// make sure blocks is a 64-byte aligned pointer, i.e., cache-line aligned
+	bf = reinterpret_cast<uint64_t *>((64ULL + reinterpret_cast<uint64_t>(buf_.get())) & ~63ULL);
+	std::fill_n(bf, num_sectors, 0);
+
+	initialized = true;
+}
+
+inline uint64_t GetMask(const hash_t hash) {
+	const uint64_t shifts = hash & SHIFT_MASK;
+	const auto shifts_8 = reinterpret_cast<const uint8_t *>(&shifts);
+
+	uint64_t mask = 0;
+
+	for (idx_t bit_idx = 8 - N_BITS; bit_idx < 8; bit_idx++) {
+		const uint8_t bit_pos = shifts_8[bit_idx];
+		mask |= (1ULL << bit_pos);
+	}
+
+	return mask;
+}
+
+void BloomFilter::InsertHashes(const Vector &hashes_v, idx_t count) const {
+	auto hashes = FlatVector::GetData<uint64_t>(hashes_v);
+	for (idx_t i = 0; i < count; i++) {
+		InsertOne(hashes[i]);
+	}
+}
+
+idx_t BloomFilter::LookupHashes(const Vector &hashes_v, SelectionVector &result_sel, const idx_t count) const {
+	D_ASSERT(hashes_v.GetVectorType() == VectorType::FLAT_VECTOR);
+	D_ASSERT(hashes_v.GetType() == LogicalType::HASH);
+
+	const auto hashes = FlatVector::GetData<uint64_t>(hashes_v);
+	idx_t found_count = 0;
+	for (idx_t i = 0; i < count; i++) {
+		result_sel.set_index(found_count, i);
+		found_count += LookupOne(hashes[i]);
+	}
+	return found_count;
+}
+
+inline void BloomFilter::InsertOne(const hash_t hash) const {
+	D_ASSERT(initialized);
+	const uint64_t bf_offset = hash & bitmask;
+	const uint64_t mask = GetMask(hash);
+	atomic<uint64_t> &slot = *reinterpret_cast<atomic<uint64_t> *>(&bf[bf_offset]);
+
+	slot.fetch_or(mask, std::memory_order_relaxed);
+}
+
+inline bool BloomFilter::LookupOne(const uint64_t hash) const {
+	D_ASSERT(initialized);
+	const uint64_t bf_offset = hash & bitmask;
+	const uint64_t mask = GetMask(hash);
+	atomic<uint64_t> &slot = *reinterpret_cast<atomic<uint64_t> *>(&bf[bf_offset]);
+	auto bf_entry = slot.load(std::memory_order_relaxed);
+
+	return (bf_entry & mask) == mask;
+}
 
 BloomFilterFunctionData::BloomFilterFunctionData(optional_ptr<BloomFilter> filter_p, bool filters_null_values_p,
                                                  const string &key_column_name_p, const LogicalType &key_type_p,
